moved ESCLient to connectors.es
