+++ .isort.cfg
@@ -1,3 +1,3 @@
 [settings]
 profile = black
-skip=lib,bin
+skip=lib,bin,include
