+++ connectors/es/__init__.py
@@ -3,6 +3,7 @@
 # or more contributor license agreements. Licensed under the Elastic License 2.0;
 # you may not use this file except in compliance with the Elastic License 2.0.
 #
+from connectors.es.client import ESClient  # NOQA
 from connectors.es.index import ESIndex  # NOQA
 from connectors.es.settings import (  # NOQA
     DEFAULT_LANGUAGE,
